backend/requirements.txt
# RaptorFlow 2.0 Backend Dependencies
# Enterprise Multi-Agent Marketing OS

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Agent Orchestration & LLMs
langgraph>=0.0.40
langchain>=0.1.10
langchain-openai>=0.0.8
langchain-community>=0.0.20
openai>=1.12.0

# Google Cloud & Vertex AI
google-cloud-aiplatform>=1.42.1
anthropic[vertex]>=0.18.0

# Database & Storage
supabase>=2.3.4
postgrest>=0.16.0
psycopg2-binary>=2.9.9

# Redis for Caching & Queue
redis>=5.0.1
hiredis>=2.3.2

# Data Validation & Settings
pydantic>=2.6.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.3

# Web Scraping & Research
beautifulsoup4>=4.12.3
lxml>=5.1.0
playwright>=1.41.2

# Social Media APIs
tweepy>=4.14.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.117.0

# Content Generation
markdown>=3.5.2
Pillow>=10.2.0
python-docx>=1.1.0

# Analytics
python-dateutil>=2.8.2

# Logging & Monitoring
structlog>=24.1.0
python-json-logger>=2.0.7

# Environment & Configuration
python-dotenv>=1.0.1

# Task Queue
celery[redis]>=5.3.6
kombu>=5.3.5

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0
pytest-mock>=3.12.0
httpx>=0.26.0  # For TestClient

# Development Tools
black>=24.1.1
ruff>=0.2.1
mypy>=1.8.0
types-redis>=4.6.0.20240106

# Vector DB (optional for semantic search)
# chromadb>=0.4.22
# pinecone-client>=3.0.2

# Rate Limiting
slowapi>=0.1.9

# CORS
python-cors>=1.0.0

# JSON Schema
jsonschema>=4.21.1

# Retry Logic
tenacity>=8.2.3

# UUID
shortuuid>=1.0.12

# Date/Time
pytz>=2024.1

